requirements.txt
discord.py==2.5.2
python-dotenv==1.1.0
playwright
jinja2
pillow